+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { ForecastDetails } from "@/components/ForecastDetails";
 import { Spinner } from "@/components/Spinner";
-import { type CompForecast } from "@/utils/calculate-wprs";
+import { type ApiResponse } from "@/utils/calculate-wprs";
 import { type NextPage } from "next";
 import Head from "next/head";
 import Link from "next/link";
@@ -8,9 +8,10 @@ import { useEffect, useState, type ChangeEvent } from "react";
 import { Footer } from "@/components/Footer";
 import { isValidUrl } from "@/utils/check-valid-url";
 import { useRouter } from "next/router";
+import { ListRankings } from "@/components/ListRankings";
 
 const Home: NextPage = () => {
-  const [compForecast, setCompForecast] = useState<CompForecast | undefined>();
+  const [compForecast, setCompForecast] = useState<ApiResponse | undefined>();
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string | undefined>();
   const [url, setUrl] = useState<string>("");
@@ -34,13 +35,13 @@ const Home: NextPage = () => {
       const response = await fetch(endpoint, options);
 
       if (response.status === 201) {
-        const resData = (await response.json()) as CompForecast;
+        const resData = (await response.json()) as ApiResponse;
 
         setCompForecast(resData);
       } else if (response.status === 429) setError("Too many requests");
       else if (response.status === 400) setError("This is not a valid link");
       else if (response.status === 204)
-        setError("Not enough confirmed pilots in this comp");
+        setError("Not enough pilots in this comp");
       else throw new Error("Ooops… something went wrong.");
     } catch (error) {
       setError("Ooops… something went wrong.");
@@ -152,7 +153,7 @@ const Home: NextPage = () => {
               <div className="text-lg">
                 WPRS:{" "}
                 <span className="font-bold text-[hsl(125,50%,56%)]">
-                  {compForecast?.WPRS[0]?.Ta3}
+                  {compForecast.confirmed.WPRS[0]?.Ta3}
                 </span>
               </div>
               <div className="text-sm text-slate-100">
@@ -171,12 +172,7 @@ const Home: NextPage = () => {
                 their CIVL rankings. The calculation will become more accurate
                 as the competition date approaches.
               </p>
-              <p className="text-sm text-slate-200">
-                Limitations: For technical reasons (lookup by name), the
-                calculation of a comp listed on civlcomps.org, swissleague.ch
-                and pwca.org is currently less accurate than on airtribune.com
-                (And takes significantly longer). I&apos;m working on it 😜
-              </p>
+              <ListRankings data={compForecast} />
             </div>
           )}
         </div>