+++ src/utils/get-swissleague-pilots.ts
@@ -1,5 +1,5 @@
 import { load } from "cheerio";
-import { lookupCivlId } from "@/utils/lookup-civl-id";
+import { getCivlId } from "@/utils/get-civl-id";
 
 export async function getSwissleaguePilots(url: string) {
   const response = await fetch(url);
@@ -29,14 +29,10 @@ export async function getSwissleaguePilots(url: string) {
     data.push(rowData);
   });
 
-  const confirmedPilots = data.filter((el) => {
-    return el.status == "Paid";
-  });
-
   const pilots = await Promise.all(
-    confirmedPilots.map(async (el) => {
+    data.map(async (el) => {
       const name = el.pilot ?? "";
-      const civlID = await lookupCivlId(name);
+      const civlID = await getCivlId(name);
 
       return {
         name,
@@ -44,9 +40,16 @@ export async function getSwissleaguePilots(url: string) {
         civlID,
         wing: el.glider,
         status: el.status,
+        confirmed: isConfirmed(el.status),
       };
     })
   );
 
   return pilots;
+}
+
+function isConfirmed(status?: string) {
+  return (
+    status?.toLowerCase() == "paid" || status?.toLowerCase() == "free entry"
+  );
 }