+++ tailwind.config.cjs
@@ -4,7 +4,7 @@ const config = {
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [require("daisyui")],
 };
 
 module.exports = config;